requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
python-dotenv
python-telegram-bot==13.15
